Tests, start of connected state, etc
As usual the help from AI went a bit overboard, so in the setup of
tests I ended up with JSDOM prematurely. But otherwise the test setup
was more painless than usual.

The tests were added because I started trying to connect the notion of
Jaxs state to the Alert's close link. I'm still trying to work out how
to easily construct a view layer, view model, data section and event
handling. I'm hoping a components library that is reusable is the
best approach since it will force these to be independent-ish.

Last thing I did was to have AI change the storybook app to always use
the same jaxs app, which seemed like the way this would work in real
life, but I'm not sure it's the right anwser for Storybook.