Adds breadcrumbs, with jaxs bugfix
The breadcrumbs story and some of the component were built via AI,
but there was a bug in Jaxs where some of the children hadn't yet
been converted to a Text template and couldn't render.

This story isn't the greatest and doesn't allow arg modification, but
it did find and motivate a fix for the bug.